Small improvements moved common logic to QueryIndex now an abstract class Removed Lazy Parsing of queries, readonly monitor uses in memory cache just like Writable, but without locking and purge executor.
